macro ver 2
This version of the macro has been slightly updated - the language is still imagej's internal language - the macro now takes a single folder with a single channel and, like the old version, saves the puncta and the masks, but also collects the min, max, mean, and standard deviation across all thresholds for a single image. This is just framework for a later update, there are some major modifications I need to make to get the stats part as I want it.