macrov3
This is pretty obnoxious, partially due to the fact that in ImageJ's version of javascript, you are not allowed to have arrays of arrays. I don't want to switch languages here because this is the most easily accessible language for my lab. We do not generally use matlab for image analysis and imagej is free and updates itself. Hugely frustrating to try and keep concise.